requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
snowflake-snowpark-python>=1.11.0
